Move messages to runntime
