+++ java/build.xml
@@ -80,7 +80,7 @@
        classpathref="compile.pure.classpath"
        includes="${pure.source}"
     />
-    <copy toDir="${classes}">
+    <copy toDir="${classes}" verbose="true">
       <fileset dir="${src}" includes="**/*.properties"/>
     </copy>
   </target>
@@ -100,6 +100,9 @@
        destdir="${runtime.classes}"
        classpathref="compile.runtime.classpath"
        />
+    <copy toDir="${runtime.classes}" verbose="true">
+      <fileset dir="${runtime.src}" includes="**/*.properties"/>
+    </copy>
     <ant dir="../scala" target="all" useNativeBasedir="true" inheritAll="no"/>
   </target>
 
@@ -128,8 +131,7 @@
     <fileset dir="${runtime.classes}"/>
     <fileset dir="${classes}"/>
     <fileset dir="${execute.classes}"/>
-    <fileset dir="${src}" excludes="com/ibm/streamsx/topology/internal/**"/>
-    <fileset dir="${src}" includes="com/ibm/streamsx/topology/internal/messages/*.properties"/>
+    <fileset dir="${src}" excludes="com/ibm/streamsx/topology/internal/** **/*.properties"/>
     </jar>
   </target>
