+++ java/src/com/ibm/streamsx/topology/builder/GraphBuilder.java
@@ -36,7 +36,7 @@ import com.ibm.streamsx.topology.internal.core.JavaFunctionalOps;
 import com.ibm.streamsx.topology.internal.core.SubmissionParameterFactory;
 import com.ibm.streamsx.topology.internal.functional.SubmissionParameter;
 import com.ibm.streamsx.topology.internal.streams.Util;
-import com.ibm.streamsx.topology.messages.Messages;
+import com.ibm.streamsx.topology.internal.messages.Messages;
 
 /**
  * Low-level graph builder. GraphBuilder provides a layer on top of
