+++ java/src/com/ibm/streamsx/topology/context/remote/RemoteContextFactory.java
@@ -3,7 +3,7 @@ package com.ibm.streamsx.topology.context.remote;
 import com.ibm.streamsx.topology.internal.context.remote.RemoteBuildAndSubmitRemoteContext;
 import com.ibm.streamsx.topology.internal.context.remote.ToolkitRemoteContext;
 import com.ibm.streamsx.topology.internal.context.remote.ZippedToolkitRemoteContext;
-import com.ibm.streamsx.topology.messages.Messages;
+import com.ibm.streamsx.topology.internal.messages.Messages;
 
 public class RemoteContextFactory {
