+++ java/src/com/ibm/streamsx/topology/generator/spl/GraphValidation.java
@@ -15,7 +15,7 @@ import java.util.Set;
 
 import com.google.gson.JsonObject;
 import com.ibm.streamsx.topology.builder.BVirtualMarker;
-import com.ibm.streamsx.topology.messages.Messages;
+import com.ibm.streamsx.topology.internal.messages.Messages;
 
 public class GraphValidation {
