+++ java/src/com/ibm/streamsx/topology/generator/spl/SPLFromFileName.java
@@ -16,7 +16,7 @@ import java.nio.charset.StandardCharsets;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 
-import com.ibm.streamsx.topology.messages.Messages;
+import com.ibm.streamsx.topology.internal.messages.Messages;
 
 /**
  * Given the path of a file containing the JSON representation of a graph,
