+++ java/src/com/ibm/streamsx/topology/internal/context/service/RemoteStreamingAnalyticsServiceStreamsContext.java
@@ -17,7 +17,7 @@ import com.ibm.streamsx.topology.internal.context.JSONStreamsContext;
 import com.ibm.streamsx.topology.internal.context.remote.RemoteBuildAndSubmitRemoteContext;
 import com.ibm.streamsx.topology.internal.context.remote.SubmissionResultsKeys;
 import com.ibm.streamsx.topology.internal.process.CompletedFuture;
-import com.ibm.streamsx.topology.messages.Messages;
+import com.ibm.streamsx.topology.internal.messages.Messages;
 
 /**
  * Context that submits the SPL to the Streaming Analytics service
