+++ java/src/com/ibm/streamsx/topology/internal/core/WindowDefinition.java
@@ -21,8 +21,8 @@ import com.ibm.streamsx.topology.function.Function;
 import com.ibm.streamsx.topology.internal.functional.FunctionalOpProperties;
 import com.ibm.streamsx.topology.internal.logic.LogicUtils;
 import com.ibm.streamsx.topology.internal.logic.ObjectUtils;
+import com.ibm.streamsx.topology.internal.messages.Messages;
 import com.ibm.streamsx.topology.logic.Identity;
-import com.ibm.streamsx.topology.messages.Messages;
 
 public class WindowDefinition<T,K> extends TopologyItem implements TWindow<T,K> {
