+++ java/src/com/ibm/streamsx/topology/internal/streams/InvokeSc.java
@@ -19,7 +19,7 @@ import java.util.logging.Logger;
 import com.google.gson.JsonObject;
 import com.ibm.streamsx.topology.context.ContextProperties;
 import com.ibm.streamsx.topology.internal.process.ProcessOutputToLogger;
-import com.ibm.streamsx.topology.messages.Messages;
+import com.ibm.streamsx.topology.internal.messages.Messages;
 
 public class InvokeSc {
