--- java/src/com/ibm/streamsx/topology/messages/.theFileNameList
@@ -1,1 +1,0 @@
-messages1.properties
