--- java/src/com/ibm/streamsx/topology/messages/messages1.properties
@@ -1,103 +1,0 @@
-# NLS_MESSAGEFORMAT_ALL
-# NLS_ENCODING=UNICODE
-# Message id range CDIST0300 - CDIST0349
-#                  CDIST3400 - CDIST3449
-
-# TRNOTE - CONSISTENT_CHECK_1 - Do not translate the word: consistent region
-CONSISTENT_CHECK_1=CDIST0300E: Functional logic operator {0} cannot be the start of a consistent region.
-
-#TRNOTE : do not translate words: unsigned, int
-ILLEGAL_TYPE_FOR_UNSIGNED=CDIST0307E: Illegal type for ''unsigned int'' ''{0}''
-UNSUPPORTED_COMPOSITE_TYPE=CDIST0308E: Unsupported composite type: ''{0}''
-REGION_CONTAIN_OPERATOR=CDIST0309E: A region must contain at least one operator.
-REGION_OVERLAPPING=CDIST0310E: Cannot have overlapping regions of different types.
-#TRNOTE : do not translate word union
-REGION_UNION=CDIST0311E: Cannot invoke ''union()'' before ending a region.
-#TRNOTE do not translate words: name, jo
-UNHANDLED_PARAM_NAME=CDIST0312E: Unhandled parameter name={0} jo={1}
-MESSAGING_KAFKA_UNABLE_TO_CREATE_FILE=CDIST0313E: Unable to create a properties file: ''{0}''
-CONSISTENT_PERIOD=CDIST0314E: period must be greater than zero:{0}
-#TRNOTE  do not translate word: drainTimeout
-CONSISTENT_DRAIN_TIMEOUT=CDIST0315E: ''drainTimeout'' must be greater than zero:{0}
-#TRNOTE  do not translate word: resetTimeout
-CONSISTENT_RESET_TIMEOUT=CDIST0316E: ''resetTimeout'' must be greater than zero:{0}
-#TRNOTE  do not translate word: maxConsecutiveResetAttempts
-CONSISTENT_MAX_CONSECUTIVE_RESET_ATTEMPTS=CDIST0317E: ''maxConsecutiveResetAttempts'' must be greater than zero:{0}
-#TRNOTE  do not translate word: SPL
-SPL_ATTRIBUTE_NOT_PRESENT=CDIST0318E: Attribute ''{0}'' not present in SPL schema ''{1}''
-#TRNOTE  do not translate word: @PrimitiveOperator
-SPL_MISSING_ANNOTATION=CDIST0319E: Missing ''@PrimitiveOperator'' for: ''{0}''
-#TRNOTE  do not translate word: SPLStream
-SPL_PARTITIONING_NOT_SUPPORTED=CDIST0320E: Partitioning is not currently supported with SPLStream.
-CONTEXT_UNKNOWN_TYPE=CDIST0321E: Unknown type: ''{0}''
-CONTEXT_JOB_SUBMISSION_FAILED=CDIST0322E: Job submission failed!
-#TRNOTE  do not translate word: config:
-CONTEXT_UNKNOWN_TYPE_FOR_CONFIG=CDIST0323E: Unknown type for config: ''{0}'' - ''{1}''
-#TRNOTE  do not translate words: streamtool, submitjob
-STREAMS_STREAMTOOL_SUBMITJOB_FAILED=CDIST0324E: ''streamtool submitjob'' failed!
-#TRNOTE  do not translate word streamtool
-STREAMS_STREAMTOOL_FAILED_TO_SUPPLY=CDIST0325E: ''streamtool'' failed to supply a job identifier!
-#TRNOTE  do not translate words: streamtool, canceljob
-STREAMS_STREAMTOOL_CANCELJOB_FAILED=CDIST0326E: ''streamtool canceljob'' failed!
-#TRNOTE  do not translate word spl-make-toolkit
-STREAMS_SPL_MAKE_TOOLKIT_FAILED=CDIST0327E: ''spl-make-toolkit'' failed!
-#TRNOTE  do not translate word Streams
-STREAMS_IS_NOT_A_STREAMS_INSTALL_DIRECTORY=CDIST0328E: ''{0}'' ''{1}'' is not a Streams install directory.
-STREAMS_ENVIRONMENT_VARIABLE_IS_NOT_SET=CDIST0329E: ''{0}'' environment variable is not set.
-#TRNOTE  do not translate word null
-STREAMS_CONFIG_ITEM_NULL=CDIST0330E: Configuration item ''{0}'' value is null
-#TRNOTE  do not translate word instanceof
-STREAMS_CONFIG_ITEM_REQUIRE=CDIST0331E: Configuration item ''{0}'' value is ''{1}'' but require instanceof ''{2}''
-CORE_AGGREGATE_PERIOD_CANNOT_BE_ZERO=CDIST0332E: Aggregate period cannot be zero."
-CORE_WINDOW_DURATION_OF_ZERO=CDIST0333E: Window duration of zero is not allowed.
-#TRNOTE  do not translate word TStream
-CORE_TSTREAM_TUPLE_TYPE=CDIST0334E: ''TStream'' tuple type cannot be published allowing filters.
-#TRNOTE  do not translate word TStream
-CORE_TSTREAM_TUPLE_GENERIC_TYPE=CDIST0335E: A ''TStream'' with a tuple type that contains a generic or unknown type cannot be published.
-CORE_INVALID_TOPIC_NAME=CDIST0336E: Invalid topic name:''{0}''
-#TRNOTE  do not translate word Routing.KEY_PARTITIONED
-#TRNOTE  do not translate words: parallel(Supplier<Integer> width, Function<T,?> keyer)
-CORE_ROUTING_KEY_PARTITIONED=CDIST0337E: ''Routing.KEY_PARTITIONED'' requires a key function. Use ''parallel(Supplier<Integer> width, Function<T,?> keyer)''.
-CORE_UNSUPPORTED_ROUTING=CDIST0338E: Unsupported routing:''{0}''
-#TRNOTE  do not translate words: keyer null
-CORE_KEYER_IS_NULL=CDIST0339E: ''keyer'' is null.
-#TRNOTE  do not translate words: Supplier: width.get(), null
-CORE_ILLEGAL_WIDTH_NULL=CDIST0340E: Illegal width Supplier: ''width.get()'' returns null.
-CORE_ILLEGAL_WIDTH_VALUE=CDIST0341E: The parallel width must be greater than or equal to 1.
-#TRNOTE  do not translate word isolate()
-CORE_ISOLATE_IN_LOW_LATENCY_REGION=CDIST0342E: ''isolate()'' is not allowed in a low latency region
-#TRNOTE  do not translate words: Placeable.isPlaceable()==false
-CORE_ILLEGAL_OPERATION_PLACEABLE=CDIST0343E: Illegal operation: Placeable.isPlaceable()==false
-#TRNOTE  do not translate words: dstDirName, absPath, null
-CORE_DIR_NAME_OR_ABS_PATH_NULL=CDIST0344E: ''dstDirName'' or ''absPath'' is null
-CORE_THIRD_PARTY_DEP=CDIST0345E: File not found. Invalid third party dependency location: ''{0}''
-CORE_DEP_DIR_INVALID=CDIST0346E: Invalid Dependency directory.
-CORE_DEP_FILE_INVALID=CDIST0347E: Not a file or directory. Invalid file dependency location: ''{0}''
-CORE_PATH_INVALID=CDIST0348E: Path not a file or directory:''{0}''
-CORE_ERROR_RESOLVING_DEP=CDIST0349E: Error resolving dependency ''{0}''
-#TRNOTE  do not translate word valueClass
-CORE_UNHANDLED_VALUECLASS=CDIST3400E: Not handled ''valueClass'' ''{0}''
-CORE_DIFFERENT_TOPOLOGIES=CDIST3401E: Different topologies: ''{0}'' and ''{1}''
-CORE_ILLEGAL_TO_COLOCATE=CDIST3402E: Illegal to colocate an isolated stream with its parent.
-#TRNOTE  do not translate word colocate()
-CORE_COLOCATE_IN_LOW_LATENCY_REGION=CDIST3403E: ''colocate()'' is not allowed in a low latency region
-EMBEDDED_TYPE_OF_PARAM_NOT_SUPPOTRTED=CDIST3404E: Type for parameter ''{0}'' is not supported:''{1}''
-EMBEDDED_TOPOLOGY_NOT_SUPPORT_EMBEDDED=CDIST3405E: Topology ''{0}'' does not support ''{1}'' mode: the topology contains non-Java operator:''{2}''
-EMBEDDED_PARAMETER_REQUIRED=CDIST3406E: Submission parameter ''{0}'' requires a value but none has been supplied
-GENERATOR_FILE_NOT_EXIST=CDIST3407E: File ''{0}'' does not exist
-GENERATOR_INTERNAL_ERROR=CDIST3408E: Internal error
-#TRNOTE  do not translate word endParallel()
-GENERATOR_CANNOT_UNION=CDIST3409E: Cannot union multiple streams before invoking ''endParallel()''
-#TRNOTE  do not translate word endParallel()
-GENERATOR_CANNOT_FANOUT=CDIST3410E: Cannot fan-out a stream before invoking ''endParallel()''
-TOPOLOGY_INVALID_TOPIC_FILTER=CDIST3411E: Invalid topic filter:''{0}''
-BUILDER_NAME_ALREADY_DEFINED=CDIST3412E: Name ''{0}'' is already defined.
-#TRNOTE  do not translate words: NULL, PARAM
-BUILDER_NULL_PARAM=CDIST3413E: NULL PARAM: ''{0}''
-#TRNOTE  do not translate word JSON
-BUILDER_ILLEGAL_JSON_OBJECT=CDIST3414E: Illegal JSON object ''{0}''
-BUILDER_TYPE_OF_PARAMER_NOT_SUPPORTED=CDIST3415E: Type for parameter ''{0}'' is not supported:''{1}''
-#TRNOTE  do not translate word createValue()
-SPL_PARAMETER_INVALID=CDIST3416E: Parameter is not from ''createValue()''
-#TRNOTE  do not translate word SPL
-STREAMS_COMPILATION_FAILED=CDIST3417E: SPL compilation failed!
