+++ java/src/com/ibm/streamsx/topology/messaging/kafka/Util.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import com.ibm.streamsx.topology.messages.Messages;
+import com.ibm.streamsx.topology.internal.messages.Messages;
 
 import com.ibm.streamsx.topology.TopologyElement;
